Move completion command out of root package
