+++ pkg/cmd/completion/completion_test.go
@@ -1,9 +1,10 @@
-package root
+package completion
 
 import (
 	"strings"
 	"testing"
 
+	"github.com/cli/cli/pkg/cmdutil"
 	"github.com/cli/cli/pkg/iostreams"
 	"github.com/google/shlex"
 	"github.com/spf13/cobra"
@@ -45,7 +46,10 @@ func TestNewCmdCompletion(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			io, _, stdout, stderr := iostreams.Test()
-			completeCmd := NewCmdCompletion(io)
+			f := &cmdutil.Factory{
+				IOStreams: io,
+			}
+			completeCmd := NewCmdCompletion(f)
 			rootCmd := &cobra.Command{Use: "gh"}
 			rootCmd.AddCommand(completeCmd)
